+++ projects/packages/forms/src/blocks/contact-form/components/form-add-step-controls/index.js
@@ -4,7 +4,7 @@ import { ToolbarGroup, MenuGroup, MenuItem, ToolbarDropdownMenu } from '@wordpre
 import { useDispatch, useSelect } from '@wordpress/data';
 import { useCallback } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
-import { store as singleStepStore } from '../../../../store/preview-store';
+import { store as singleStepStore } from '../../../../store/form-step-preview';
 import useContainerId from '../../../shared/hooks/use-container-id';
 import useFormSteps from '../../../shared/hooks/use-form-steps';
