+++ projects/packages/forms/src/blocks/form-step-container/icon.js
