--- projects/packages/forms/src/blocks/form-step-navigation/style.scss
@@ -1,52 +1,0 @@
-// Frontend styles for the jetpack-form-step-navigation block
-
-.jetpack-form-step-navigation {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	gap: 1rem;
-	margin: 1.5rem 0;
-
-	&__button {
-		display: inline-flex;
-		align-items: center;
-		justify-content: center;
-		padding: 0.75rem 1.5rem;
-		border-radius: 4px;
-		font-weight: 500;
-		text-decoration: none;
-		transition: all 0.2s ease-in-out;
-		cursor: pointer;
-		border: 1px solid transparent;
-
-		&--previous {
-			background-color: var( --wp--preset--color--background, #fff );
-			color: var( --wp--preset--color--primary, #1e1e1e );
-			border-color: var( --wp--preset--color--primary, #1e1e1e );
-
-			&:hover {
-				background-color: var( --wp--preset--color--primary, #1e1e1e );
-				color: var( --wp--preset--color--background, #fff );
-			}
-		}
-
-		&--next {
-			background-color: var( --wp--preset--color--primary, #1e1e1e );
-			color: var( --wp--preset--color--background, #fff );
-
-			&:hover {
-				background-color: var( --wp--preset--color--secondary, #2c3338 );
-			}
-		}
-
-		&:disabled {
-			opacity: 0.5;
-			cursor: not-allowed;
-		}
-	}
-
-	&__step-indicator {
-		font-size: 0.875rem;
-		color: var( --wp--preset--color--secondary, #2c3338 );
-	}
-}
