+++ projects/packages/forms/src/blocks/form-step/icon.js
@@ -1,9 +1,8 @@
-import { SVG, Path } from '@wordpress/components';
+import { Path } from '@wordpress/components';
+import renderMaterialIcon from '../shared/components/render-material-icon';
 
-export default function StepIcon( props ) {
-	return (
-		<SVG width={ 24 } height={ 24 } viewBox="0 0 24 24" { ...props }>
-			<Path d="M17.5 9V6C17.5 5.46957 17.2893 4.96086 16.9142 4.58579C16.5391 4.21071 16.0304 4 15.5 4H8.5C7.96957 4 7.46086 4.21071 7.08579 4.58579C6.71071 4.96086 6.5 5.46957 6.5 6V9H8V6C8 5.86739 8.05268 5.74021 8.14645 5.64645C8.24021 5.55268 8.36739 5.5 8.5 5.5H15.5C15.6326 5.5 15.7598 5.55268 15.8536 5.64645C15.9473 5.74021 16 5.86739 16 6V9H17.5ZM17.5 15.5V18C17.5 18.5304 17.2893 19.0391 16.9142 19.4142C16.5391 19.7893 16.0304 20 15.5 20H8.5C7.96957 20 7.46086 19.7893 7.08579 19.4142C6.71071 19.0391 6.5 18.5304 6.5 18V15.5H8V18C8 18.1326 8.05268 18.2598 8.14645 18.3536C8.24021 18.4473 8.36739 18.5 8.5 18.5H15.5C15.6326 18.5 15.7598 18.4473 15.8536 18.3536C15.9473 18.2598 16 18.1326 16 18V15.5H17.5ZM4 13H20V11.5H4V13Z" />
-		</SVG>
-	);
-}
+const StepIcon = renderMaterialIcon(
+	<Path d="M17.5 9V6C17.5 5.46957 17.2893 4.96086 16.9142 4.58579C16.5391 4.21071 16.0304 4 15.5 4H8.5C7.96957 4 7.46086 4.21071 7.08579 4.58579C6.71071 4.96086 6.5 5.46957 6.5 6V9H8V6C8 5.86739 8.05268 5.74021 8.14645 5.64645C8.24021 5.55268 8.36739 5.5 8.5 5.5H15.5C15.6326 5.5 15.7598 5.55268 15.8536 5.64645C15.9473 5.74021 16 5.86739 16 6V9H17.5ZM17.5 15.5V18C17.5 18.5304 17.2893 19.0391 16.9142 19.4142C16.5391 19.7893 16.0304 20 15.5 20H8.5C7.96957 20 7.46086 19.7893 7.08579 19.4142C6.71071 19.0391 6.5 18.5304 6.5 18V15.5H8V18C8 18.1326 8.05268 18.2598 8.14645 18.3536C8.24021 18.4473 8.36739 18.5 8.5 18.5H15.5C15.6326 18.5 15.7598 18.4473 15.8536 18.3536C15.9473 18.2598 16 18.1326 16 18V15.5H17.5ZM4 13H20V11.5H4V13Z" />
+);
+
+export default StepIcon;
