--- projects/packages/forms/src/blocks/form/style.scss
@@ -1,101 +1,0 @@
-// Frontend styles for the jetpack-form block
-
-.jetpack-form {
-	position: relative;
-	width: 100%;
-	max-width: 100%;
-	margin: 0 auto;
-
-	// Form container
-	&__container {
-		position: relative;
-		width: 100%;
-		overflow: hidden;
-	}
-
-	// Form steps wrapper
-	&__steps {
-		position: relative;
-		width: 100%;
-	}
-
-	// Form fields
-	&__field {
-		margin-bottom: 1.5rem;
-
-		&--error {
-
-			.jetpack-form__field-label {
-				color: var( --wp--preset--color--error, #d63638 );
-			}
-
-			.jetpack-form__field-input {
-				border-color: var( --wp--preset--color--error, #d63638 );
-			}
-		}
-	}
-
-	&__field-label {
-		display: block;
-		margin-bottom: 0.5rem;
-		font-weight: 500;
-	}
-
-	&__field-input {
-		width: 100%;
-		padding: 0.75rem;
-		border: 1px solid var( --wp--preset--color--border, #ddd );
-		border-radius: 4px;
-		transition: border-color 0.2s ease-in-out;
-
-		&:focus {
-			outline: none;
-			border-color: var( --wp--preset--color--primary, #1e1e1e );
-		}
-	}
-
-	&__field-error {
-		display: none;
-		margin-top: 0.5rem;
-		color: var( --wp--preset--color--error, #d63638 );
-		font-size: 0.875rem;
-
-		.jetpack-form__field--error & {
-			display: block;
-		}
-	}
-
-	// Form navigation
-	&__navigation {
-		margin-top: 2rem;
-	}
-
-	// Form progress
-	&__progress {
-		margin-bottom: 2rem;
-	}
-
-	// Form submit button
-	&__submit {
-		display: inline-flex;
-		align-items: center;
-		justify-content: center;
-		padding: 0.75rem 1.5rem;
-		background-color: var( --wp--preset--color--primary, #1e1e1e );
-		color: var( --wp--preset--color--background, #fff );
-		border: none;
-		border-radius: 4px;
-		font-weight: 500;
-		cursor: pointer;
-		transition: background-color 0.2s ease-in-out;
-
-		&:hover {
-			background-color: var( --wp--preset--color--secondary, #2c3338 );
-		}
-
-		&:disabled {
-			opacity: 0.5;
-			cursor: not-allowed;
-		}
-	}
-}
