+++ projects/packages/forms/src/blocks/shared/hooks/use-step-navigation.js
@@ -1,7 +1,7 @@
 import { store as blockEditorStore } from '@wordpress/block-editor';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { useCallback } from '@wordpress/element';
-import { store as singleStepStore } from '../../../store/preview-store';
+import { store as singleStepStore } from '../../../store/form-step-preview';
 import useFormSteps from './use-form-steps';
 
 /**
